+++ lib/que.rb
@@ -16,9 +16,38 @@ module Que
     JSON_MODULE = JSON
   end
 
+  HASH_DEFAULT_PROC = proc { |hash, key| hash[key.to_s] if Symbol === key }
+
+  INDIFFERENTIATOR = proc do |object|
+    case object
+    when Array
+      object.each(&INDIFFERENTIATOR)
+    when Hash
+      object.default_proc = HASH_DEFAULT_PROC
+      object.each { |key, value| object[key] = INDIFFERENTIATOR.call(value) }
+      object
+    else
+      object
+    end
+  end
+
+  SYMBOLIZER = proc do |object|
+    case object
+    when Hash
+      object.keys.each do |key|
+        object[key.to_sym] = SYMBOLIZER.call(object.delete(key))
+      end
+      object
+    when Array
+      object.map! { |e| SYMBOLIZER.call(e) }
+    else
+      object
+    end
+  end
+
   class << self
     attr_accessor :error_handler
-    attr_writer :logger, :adapter, :log_formatter, :disable_prepared_statements
+    attr_writer :logger, :adapter, :log_formatter, :disable_prepared_statements, :json_converter
 
     def connection=(connection)
       self.adapter =
@@ -124,6 +153,10 @@ module Que
           end
         end
       end
+    end
+
+    def json_converter
+      @json_converter ||= INDIFFERENTIATOR
     end
 
     # Copy some of the Worker class' config methods here for convenience.